stimuli: midi_proc_diff_unbounded now keeps the current value in "last"
This means that changing the control variable will not affect the
value written when the next MIDI message arrives. This is to make
diff -> unbounded behaviour consistent with the rest.

+++ src/renderer/stimuli.h
@@ -42,7 +42,7 @@ struct stim_regs {
 struct s_midi_ctrl {
 	void (*proc)(struct s_midi_ctrl *sct, int value);
 	struct stim_regs regs;
-	uint8_t last;		/* for midi_proc_diff_* */
+	long last;		/* for midi_proc_diff_* */
 };
 
 struct s_midi_chan {
